Historical record of LAMPS
As prepared by Paul Hardy. This should be seen as the first of the
more definitive code commits. The ones prior to this are just fragments
of code detected as stray saved files or comparisons to other versions
of the code and are unlikely to be complete enough to be useful.